requirements.txt
django==3.1
djangorestframework==3.11.1
pillow==8.0.1
pytils
python-decouple
django-filter
django-rest-swagger
drf-yasg==1.17.1